+++ Calculator/Calculator.Tests/PowCalculationTest.cs
@@ -9,7 +9,7 @@ namespace Calculator.Tests
         [TestCase(2, 4, 16)]
         [TestCase(12, 2, 144)]
         [TestCase(16, 2, 256)]
-        public void CalculateTest(int argument, int secondArgument, int result)
+        public void CalculateTest(double argument, double secondArgument, double result)
         {
             var calculator = new Pow();
             var testResult = calculator.Calculate(argument, secondArgument);
